added gitignore; read and write to binary sorted

+++ lib/blob/types.h
@@ -2,7 +2,7 @@
 
 namespace jp{
 
-typedef int    id_t;
+typedef uint8_t    id_t;
 typedef double data_t;
